+++ packages/elements/src/MatchersCache.spec.ts
@@ -15,6 +15,8 @@ describe("MatchersCache", () => {
       includePaths: ["src/**"],
       ignorePaths: ["dist/**"],
       cache: true,
+      multiMatch: true,
+      elementDescriptorsPriority: "last",
       rootPath: "/root",
       flagAsExternal: {
         inNodeModules: true,
@@ -39,7 +41,7 @@ describe("MatchersCache", () => {
     const key = matchersCache.getKey({ config, elementDescriptors });
 
     expect(key).toBe(
-      "true|src/**|dist/**|true|/root|true|false|true|@external/*|:|component|ui|src/components/*.tsx|src|file|name|baseName"
+      "true|src/**|dist/**|true|true|last|/root|true|false|true|@external/*|:|component|ui|src/components/*.tsx|src|file|name|baseName"
     );
   });
 });